Finished splitting RecordingView into two separate components for audio and Video.
